fix compilation and handle format
