+++ tests/BinkyLabs.OpenApi.Overlays.Cli.Tests/BinkyLabs.OpenApi.Overlays.Cli.Tests.csproj
@@ -5,8 +5,6 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
-    <SignAssembly>True</SignAssembly>
-    <AssemblyOriginatorKeyFile>../../keyfile.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
 
   <ItemGroup>
@@ -22,14 +20,13 @@
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
   </ItemGroup>
-
+  
   <ItemGroup>
-    <Using Include="Xunit" />
+    <ProjectReference Include="..\..\src\tool\BinkyLabs.OpenApi.Overlays.Cli.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\lib\BinkyLabs.OpenApi.Overlays.csproj" />
-    <ProjectReference Include="..\..\src\tool\BinkyLabs.OpenApi.Overlays.Cli.csproj" />
+    <Using Include="Xunit" />
   </ItemGroup>
 
 </Project>
